requirements.txt
Mastodon.py
pandas
matplotlib
fastapi
deta

